FrmUrunler Revize Bugfix

+++ Ticari_Otomasyon/FrmUrunler.cs
@@ -31,35 +31,34 @@ namespace Ticari_Otomasyon
         private void FrmUrunler_Load(object sender, EventArgs e)
         {
             listele();
+            temizle();
         }
 
         private void BtnKaydet_Click(object sender, EventArgs e)
         {
-            //Verileri kaydet
-            SqlCommand komut =
-                new SqlCommand("insert into TBL_URUNLER (URUNAD,MARKA,MODEL,YIL,ADET,ALISFIYAT,SATISFIYAT,DETAY) values (@p1,@p2,@p3,@p4,@p5,@p6,@p7,@p8)", bgl.baglanti());
-            komut.Parameters.AddWithValue("@p1", TxtAd.Text);
-            komut.Parameters.AddWithValue("@p2", TxtMarka.Text);
-            komut.Parameters.AddWithValue("@p3", TxtModel.Text);
-            komut.Parameters.AddWithValue("@p4", MskYil.Text);
-            komut.Parameters.AddWithValue("@p5", int.Parse(NudAdet.Value.ToString()));
-            komut.Parameters.AddWithValue("@p6", decimal.Parse(TxtAlis.Text));
-            komut.Parameters.AddWithValue("@p7", decimal.Parse(TxtSatis.Text));
-            komut.Parameters.AddWithValue("@p8", RchDetay.Text);
-            komut.ExecuteNonQuery();
-            bgl.baglanti().Close();
-            MessageBox.Show("Ürün sisteme eklendi", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            FrmYeniUrun yeniurunformu = new FrmYeniUrun();
+            yeniurunformu.ShowDialog(); // Modally form açılır
+
+            // Yeni kullanıcı eklenip form kapandığında listeyi güncelle
             listele();
+            temizle();
         }
 
         private void BtnSil_Click(object sender, EventArgs e)
         {
-            SqlCommand komutsil = new SqlCommand("Delete From TBL_URUNLER where ID=@p1", bgl.baglanti());
-            komutsil.Parameters.AddWithValue("@p1", TxtID.Text);
-            komutsil.ExecuteNonQuery();
-            bgl.baglanti().Close();
-            MessageBox.Show("Ürün silindi", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Error);
-            listele();
+            if (TxtID.Text == "")
+            {
+                MessageBox.Show("Seçim Yapmadınız", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+            else
+            {
+                SqlCommand komutsil = new SqlCommand("Delete From TBL_URUNLER where ID=@p1", bgl.baglanti());
+                komutsil.Parameters.AddWithValue("@p1", TxtID.Text);
+                komutsil.ExecuteNonQuery();
+                bgl.baglanti().Close();
+                MessageBox.Show("Ürün silindi", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                listele();
+            }
         }
 
         private void gridView1_FocusedRowChanged(object sender, DevExpress.XtraGrid.Views.Base.FocusedRowChangedEventArgs e)
@@ -78,7 +77,13 @@ namespace Ticari_Otomasyon
 
         private void BtnGuncelle_Click(object sender, EventArgs e)
         {
-            SqlCommand komut = new SqlCommand("update TBL_URUNLER set URUNAD=@P1,MARKA=@P2,MODEL=@P3,YIL=@P4,ADET=@P5,ALISFIYAT=@P6,SATISFIYAT=@P7,DETAY=@P8 where ID=@P9", bgl.baglanti());
+            if (TxtID.Text == "")
+            {
+                MessageBox.Show("Seçim Yapmadınız", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+            else
+            {
+                SqlCommand komut = new SqlCommand("update TBL_URUNLER set URUNAD=@P1,MARKA=@P2,MODEL=@P3,YIL=@P4,ADET=@P5,ALISFIYAT=@P6,SATISFIYAT=@P7,DETAY=@P8 where ID=@P9", bgl.baglanti());
             komut.Parameters.AddWithValue("@p1", TxtAd.Text);
             komut.Parameters.AddWithValue("@p2", TxtMarka.Text);
             komut.Parameters.AddWithValue("@p3", TxtModel.Text);
@@ -92,6 +97,24 @@ namespace Ticari_Otomasyon
             bgl.baglanti().Close();
             MessageBox.Show("Ürün bilgisi güncellendi", "Bilgi", MessageBoxButtons.OK, MessageBoxIcon.Warning);
             listele();
+            }
+        }
+        void temizle()
+        {
+            TxtID.Text = "";
+            TxtAd.Text = "";
+            TxtMarka.Text = "";
+            TxtModel.Text = "";
+            MskYil.Text = "";
+            NudAdet.Text = "";
+            TxtAlis.Text = "";
+            TxtSatis.Text = "";
+            RchDetay.Text = "";
+        }
+
+        private void BtnTemizle_Click(object sender, EventArgs e)
+        {
+            temizle();
         }
     }
 }
